+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - License GPL v2
+- FreeRTOS instead of cooperative loop
 
 ### Removed